Change 'Program Files' to 'Program Files (x86)'
This is what the path is on my Win7 system, and what's mentioned in #2. It also seems like most people are using 64-bit Windows these days, or whatever it is that triggers the x86 variant of the Program Files directory, so I thought it was better to just change it rather than add special instructions for 64-bit Windows.